Fixed data type errors

+++ src/Ground-Terminal-System/ChecksumValidator.cs
@@ -3,10 +3,10 @@ namespace GroundTerminalSystem
     public class ChecksumValidator
     {
 
-        public bool Validate(float altitude, float pitch, float bank, int receivedChecksum)
+        public bool Validate(double altitude, double pitch, double bank, int receivedChecksum)
         {
             // Checksum = (ALT + Pitch + Bank) / 3
-            float calculation = (altitude + pitch + bank) / 3;
+            double calculation = ((altitude + pitch + bank) / 3);
 
             int calculatedChecksum = (int)calculation;
